t_joueur_init(3) prend l'id en int, non en short

+++ moteur/moteur.h
@@ -71,7 +71,7 @@ typedef struct {
     void t_case_init(t_case* t_case, int x, int y); // initialisation de la structure
 
 // STRUCT T_JOUEUR (dans struct_case_joueur.c)
-    void t_joueur_init(t_joueur *j, short id, bool ia); // initialise le joueur selon le nombre denjoueur initialisé auparavant
+    void t_joueur_init(t_joueur *j, int id, bool ia); // initialise le joueur selon le nombre denjoueur initialisé auparavant
 
 // STRUCT T_JEU (dans struct_jeu.c)
     t_jeu* t_jeu_init(short nbjoueurs, short nbIA); // allocation et initialisation de la structure. Retourne l'adresse de la structure ou NULL
